Initial check in of SQL network trace script
This will take a network trace when connecting to a SQL instance or AG Listener.  The goal of this function is to minimize the network trace size. The script will start the network trace (netsh trace start), and immediately try to connect with System.Data.SqlClient, then stop the traces as soon as it fails or succeeds. 
It should be run from the client system by someone that is an Admin. 
The default is a single sided client network trace, unless TraceType 'TwoSided' is specified.
For a two sided trace to work, you will need to be an Admin on both client and the SQL system and a valid logPath that exists on both systems should be provided.