bugfix: limit SQL instance returned to match
There was a bug here from Microsoft's end that would return a list of SQL instances if they are partial matches. For instance, this command when ran to retrieve database 'Test' will return a list if you have multiple databases called 'Test', 'TestTemp', 'TestDev'. The only way I found to fix this issue and keep it running on schedule is if we filter the returned result down again to only be the instance which ends in ";$BkupToRestore" (hence ensuring no prefix AND no postfix) . This worked for my case and I don't envision it would be a breaking change, just a bugfix.